+++ packages/extension-chakra-storefront/src/pages/product-list/page-metadata.jsx
